+++ semcore/inline-edit/__tests__/inline-edit.axe-test.tsx
@@ -25,4 +25,27 @@ test.describe('InlineEdit', () => {
       expect(violations).toEqual([]);
     }
   });
+
+  test('Editable tag', async ({ page }) => {
+    const standPath = 'stories/components/inline-edit/docs/examples/editable_tag.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+
+    // base check
+    {
+      const violations = await getAccessibilityViolations({ page });
+
+      expect(violations).toEqual([]);
+    }
+
+    // focused editor check
+    {
+      await page.keyboard.press('Tab');
+
+      const violations = await getAccessibilityViolations({ page });
+
+      expect(violations).toEqual([]);
+    }
+  });
 });