+++ stories/components/link/docs/examples/link_addon.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Link from 'intergalactic/link';
-import CheckM from 'intergalactic/icon/Check/m';
-import ChevronRightM from 'intergalactic/icon/ChevronRight/m';
-import FormatText from 'intergalactic/format-text';
+import Link from '@semcore/link';
+import CheckM from '@semcore/icon/Check/m';
+import ChevronRightM from '@semcore/icon/ChevronRight/m';
+import FormatText from '@semcore/format-text';
 
 const Demo = () => {
   return (