+++ stories/patterns/filters/advanced-filters/docs/index.stories.tsx
@@ -2,6 +2,9 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import FiltersWithConditionsEx from './examples/filters-with-filter-conditions';
 
+import { AdvancedFiltersTest } from './__tests__/filters-with-filter-conditions.test';
+import { playWrapper } from '../../../../utils/playWrapper';
+
 const meta: Meta = {
   title: 'Patterns/Filters/FiltersWithConditions',
 };
@@ -10,4 +13,5 @@ export default meta;
 
 export const FiltersWithConditions: StoryObj = {
   render: FiltersWithConditionsEx,
+  play: playWrapper(AdvancedFiltersTest),
 };