+++ website/docs/components/date-picker/examples/date_range_comparator_advanced_use.tsx
@@ -47,13 +47,19 @@ const Demo = () => {
           <DateRangeComparator.Footer>
             Place for a hint, useful message or controls.
           </DateRangeComparator.Footer>
-          <Notice theme='warning' style={{ borderTopLeftRadius: 0, borderTopRightRadius: 0 }}>
+          <Notice
+            aria-labelledby='comparator-notice-title'
+            theme='warning'
+            style={{ borderTopLeftRadius: 0, borderTopRightRadius: 0 }}
+          >
             <Notice.Content>
               <Flex mb={1}>
                 <Notice.Label>
                   <WarningIcon />
                 </Notice.Label>
-                <Text bold>Notice heading</Text>
+                <Text bold id='comparator-notice-title'>
+                  Warning
+                </Text>
               </Flex>
               If you change your location, all previously collected data for this article will be
               lost.