+++ website/docs/components/notice/examples/noticesmart.tsx
@@ -16,7 +16,7 @@ const Demo = () => {
     <Flex direction='column' gap={3}>
       <NoticeSmart
         label={<QuestionAltM />}
-        aria-label='New tool was launched'
+        aria-label='New tool announcement'
         closable
         onClose={() => setFirstHidden(true)}
         hidden={firstHidden}
@@ -40,6 +40,7 @@ const Demo = () => {
         </Text>
       </NoticeSmart>
       <NoticeSmart
+        aria-label='New feature announcement'
         theme='success'
         label={<ThumbUpM />}
         closable
@@ -59,7 +60,7 @@ const Demo = () => {
           experience with unparalleled innovation and functionality.
         </Text>
       </NoticeSmart>
-      <NoticeSmart theme='warning' label={<WarningM />}>
+      <NoticeSmart theme='warning' label={<WarningM />} aria-label='Maintenance notice'>
         <Text tag='div' my={1}>
           The Link Building tool is under maintenance. Starting new campaigns is temporarily
           unavailable, but you can continue working with your existing Link Building campaigns.